+++ src/main/java/com/niuqu/chatbubble/chat/notification/MentionNotificationBanner.java
@@ -198,8 +198,8 @@ public class MentionNotificationBanner {
         int textW = current.textW;
         int bannerW = current.bannerW;
         int bannerH = current.bannerH;
-        int x = (screenW - bannerW) / 2;
-        int y = (int) ((-bannerH) + slide * bannerH);
+        int x = (screenW - bannerW) / 2 + ChatBubbleClientSetup.config().bannerOffsetX();
+        int y = (int) ((-bannerH) + slide * bannerH) + ChatBubbleClientSetup.config().bannerOffsetY();
 
         int shadowAlpha = (int) (0x30 * alpha);
         int shadowColor = (shadowAlpha << 24);